rework performance factory
