+++ performance_test/CMakeLists.txt
@@ -26,7 +26,6 @@ set(LIBRARY_SRC
   src/resource_usage_logger.cpp
   src/system.cpp
   src/tracker.cpp
-  src/utils/cli_args.cpp
   src/utils/cli_options.cpp
   src/utils/fork_process.cpp
   src/utils/names_utilities.cpp
