+++ performance_test/include/performance_test/executors.hpp
@@ -10,39 +10,42 @@
 #pragma once
 
 #include <memory>
+#include <ostream>
 #include <string>
 
 #include "rclcpp/rclcpp.hpp"
 
-namespace performance_test {
+namespace performance_test
+{
 
 struct NamedExecutor
 {
-    std::shared_ptr<rclcpp::Executor> executor;
-    std::string name;
+  std::shared_ptr<rclcpp::Executor> executor;
+  std::string name;
 };
 
 enum ExecutorType
 {
-    SINGLE_THREADED_EXECUTOR = 1,
-    STATIC_SINGLE_THREADED_EXECUTOR = 2,
+  SINGLE_THREADED_EXECUTOR = 1,
+  STATIC_SINGLE_THREADED_EXECUTOR = 2,
 };
 
-std::ostream &operator<<(std::ostream &os, const ExecutorType &t) {
-    std::string executor_name;
-    switch (t) {
-        case SINGLE_THREADED_EXECUTOR:
-            executor_name = "SingleThreadedExecutor";
-            break;
-        case STATIC_SINGLE_THREADED_EXECUTOR:
-            executor_name = "StaticSingleThreadedExecutor";
-            break;
-        default:
-            executor_name = "Unknown ExecutorType";
-            break;
-    }
+inline std::ostream &operator<<(std::ostream &os, const ExecutorType &t)
+{
+  std::string executor_name;
+  switch (t) {
+    case SINGLE_THREADED_EXECUTOR:
+      executor_name = "SingleThreadedExecutor";
+      break;
+    case STATIC_SINGLE_THREADED_EXECUTOR:
+      executor_name = "StaticSingleThreadedExecutor";
+      break;
+    default:
+      executor_name = "Unknown ExecutorType";
+      break;
+  }
 
-    return os << executor_name;
+  return os << executor_name;
 }
 
 }
