+++ performance_test/include/performance_test/system.hpp
@@ -9,14 +9,12 @@
 
 #pragma once
 
-#include <vector>
-#include <fstream>
+#include <chrono>
 #include <map>
-#include <pthread.h>
+#include <memory>
+#include <vector>
 
-#include "performance_test/utils/names_utilities.hpp"
-#include "performance_test/utils/stat_logger.hpp"
-#include "performance_test/node.hpp"
+#include "performance_test/performance_node_base.hpp"
 #include "performance_test/executors.hpp"
 #include "performance_test/events_logger.hpp"
 
@@ -25,7 +23,6 @@ namespace performance_test {
 class System
 {
 public:
-
   System(ExecutorType executor);
 
   template<typename NodeT>
