--- performance_test/include/performance_test/utils/cli_args.hpp
@@ -1,19 +1,0 @@
-/* Software License Agreement (BSD License)
- *
- *  Copyright (c) 2019, iRobot ROS
- *  All rights reserved.
- *
- *  This file is part of ros2-performance, which is released under BSD-3-Clause.
- *  You may use, distribute and modify this code under the BSD-3-Clause license.
- */
-
-#pragma once
-
-#include <vector>
-
-namespace performance_test
-{
-
-std::vector<char *> get_non_ros_args(int argc, char** argv);
-
-}
