+++ performance_test/include/performance_test/utils/cli_options.hpp
@@ -35,4 +35,6 @@ public:
   performance_test::Tracker::TrackingOptions tracking_options;
 };
 
+std::ostream &operator<<(std::ostream &os, const Options &options);
+
 }
