+++ performance_test/src/system.cpp
@@ -15,7 +15,7 @@
 #include "performance_test/system.hpp"
 #include "performance_test/utils/names_utilities.hpp"
 #include "performance_test/utils/stat_logger.hpp"
-#include "performance_test/node.hpp"
+#include "performance_test/performance_node_base.hpp"
 #include "performance_test/executors.hpp"
 #include "performance_test/events_logger.hpp"
