+++ performance_test/test/test_system.cpp
@@ -10,7 +10,7 @@
 #include <gtest/gtest.h>
 
 #include "performance_test/system.hpp"
-#include "performance_test/node.hpp"
+#include "performance_test/performance_node.hpp"
 #include "performance_test_msgs/msg/sample.hpp"
 #include "performance_test_msgs/srv/sample.hpp"
