+++ performance_test_plugin_cmake/cmake/generate_factory_plugin.cmake
@@ -1,32 +1,32 @@
 # This is the path where this file is located
 set(GENERATE_FACTORY_PLUGIN__INTERNAL_DIR__ ${CMAKE_CURRENT_LIST_DIR} CACHE INTERNAL "")
 
-
 function(generate_factory_plugin argMSGS argSRVS)
 
   find_package(rclcpp REQUIRED)
   find_package(performance_test REQUIRED)
+  find_package(Python3 COMPONENTS Interpreter REQUIRED)
 
-  set (LIBRARY_DEPENDENCIES
+  set(LIBRARY_DEPENDENCIES
     rclcpp
     performance_test
   )
 
-  set (CUSTOM_TARGET_NAME ${PROJECT_NAME}_implementation.cpp)
-  set (CUSTOM_TARGET_PATH ${PROJECT_BINARY_DIR}/generated/${CUSTOM_TARGET_NAME})
+  set(CUSTOM_TARGET_NAME ${PROJECT_NAME}_implementation.cpp)
+  set(CUSTOM_TARGET_PATH ${PROJECT_BINARY_DIR}/generated/${CUSTOM_TARGET_NAME})
 
   add_custom_command(
-      OUTPUT ${CUSTOM_TARGET_PATH}
-      COMMAND python3
-        ${GENERATE_FACTORY_PLUGIN__INTERNAL_DIR__}/generate_script.py
-        ${CUSTOM_TARGET_PATH}
-        --package ${PROJECT_NAME}
-        --msg ${argMSGS}
-        --srv ${argSRVS}
-      DEPENDS ${PROJECT_NAME}
+    OUTPUT ${CUSTOM_TARGET_PATH}
+    COMMAND ${Python3_EXECUTABLE}
+      ${GENERATE_FACTORY_PLUGIN__INTERNAL_DIR__}/generate_script.py
+      ${CUSTOM_TARGET_PATH}
+      --package ${PROJECT_NAME}
+      --msg ${argMSGS}
+      --srv ${argSRVS}
+    DEPENDS ${PROJECT_NAME}
   )
 
-  set (LIBRARY_NAME ${PROJECT_NAME}_implementation)
+  set(LIBRARY_NAME ${PROJECT_NAME}_implementation)
   add_library(${LIBRARY_NAME} SHARED
     ${CUSTOM_TARGET_PATH}
   )
