+++ performance_test_plugin_cmake/cmake/generate_script.py
@@ -13,7 +13,7 @@ import argparse
 import os
 import re
 import sys
-
+import textwrap
 
 def get_interface_name_from_path(interface_path):
   '''
@@ -30,7 +30,7 @@ def get_interface_name_from_path(interface_path):
 
 def get_lowercased_name(interface_name):
   '''
-  interface_name has the PoseWithCovariance
+  interface_name has the form PoseWithCovariance
   the lowercased version is pose_with_covariance
   '''
 
@@ -77,11 +77,17 @@ def get_namespaced_cpp_class_name(interface_name, package, interface_type):
 
 def get_include_paths(msgs, srvs, package):
   '''
-  create include definitions for all the messages and services
+  Add all necessary include directives
   '''
 
+  content = """
+  #include <map>
+  #include <memory>
+  #include <string>
+  #include <rclcpp/rclcpp.hpp>
+  #include "performance_test/performance_node_base.hpp"
 
-  content = "\n"
+  """
 
   for msg_name in msgs:
     statement = get_cpp_include_statement(msg_name, package, "msg")
@@ -93,26 +99,17 @@ def get_include_paths(msgs, srvs, package):
 
   return content
 
-
-def get_sub_factory(msgs, package, node_type):
+def get_sub_factory(msgs, package):
 
   if len(msgs) == 0:
     return ""
 
-  if node_type == "rclcpp::Node":
-    content = """
+  content = """
 
-    extern "C" void add_subscriber_impl(
-    """
-  elif node_type == "rclcpp_lifecycle::LifecycleNode":
-    content = """
+  extern "C" void add_subscriber_impl(
+  """
 
-    extern "C" void add_subscriber_impl_lifecycle(
-    """
-  else:
-    return ""
-
-  performance_node = "performance_test::PerformanceNode<" + node_type + ">"
+  performance_node = "performance_test::PerformanceNodeBase"
   content += "\n std::shared_ptr<" + performance_node + "> n,"
 
   content += """
@@ -158,26 +155,17 @@ def get_sub_factory(msgs, package, node_type):
 
   return content
 
-
-def get_pub_factory(msgs, package, node_type):
+def get_pub_factory(msgs, package):
 
   if len(msgs) == 0:
     return ""
 
-  if node_type == "rclcpp::Node":
-    content = """
+  content = """
 
-    extern "C" void add_publisher_impl(
-    """
-  elif node_type == "rclcpp_lifecycle::LifecycleNode":
-    content = """
+  extern "C" void add_publisher_impl(
+  """
 
-    extern "C" void add_publisher_impl_lifecycle(
-    """
-  else:
-    return ""
-
-  performance_node = "performance_test::PerformanceNode<" + node_type + ">"
+  performance_node = "performance_test::PerformanceNodeBase"
   content += "\n std::shared_ptr<" + performance_node + "> n,"
 
   content += """
@@ -224,26 +212,17 @@ def get_pub_factory(msgs, package, node_type):
 
   return content
 
-
-def get_server_factory(srvs, package, node_type):
+def get_server_factory(srvs, package):
 
   if len(srvs) == 0:
     return ""
 
-  if node_type == "rclcpp::Node":
-    content = """
+  content = """
 
-    extern "C" void add_server_impl(
-    """
-  elif node_type == "rclcpp_lifecycle::LifecycleNode":
-    content = """
+  extern "C" void add_server_impl(
+  """
 
-    extern "C" void add_server_impl_lifecycle(
-    """
-  else:
-    return ""
-
-  performance_node = "performance_test::PerformanceNode<" + node_type + ">"
+  performance_node = "performance_test::PerformanceNodeBase"
   content += "\n std::shared_ptr<" + performance_node + "> n,"
 
   content += """
@@ -287,26 +266,17 @@ def get_server_factory(srvs, package, node_type):
 
   return content
 
-
-def get_client_factory(srvs, package, node_type):
+def get_client_factory(srvs, package):
 
   if len(srvs) == 0:
     return ""
 
-  if node_type == "rclcpp::Node":
-    content = """
+  content = """
 
-    extern "C" void add_client_impl(
-    """
-  elif node_type == "rclcpp_lifecycle::LifecycleNode":
-    content = """
+  extern "C" void add_client_impl(
+  """
 
-    extern "C" void add_client_impl_lifecycle(
-    """
-  else:
-    return ""
-
-  performance_node = "performance_test::PerformanceNode<" + node_type + ">"
+  performance_node = "performance_test::PerformanceNodeBase"
   content += "\n std::shared_ptr<" + performance_node + "> n,"
 
   content += """
@@ -350,7 +320,6 @@ def get_client_factory(srvs, package, node_type):
 
   return content
 
-
 def main():
 
   parser = argparse.ArgumentParser(description='Python script for generating interfaces implementation')
@@ -380,25 +349,15 @@ def main():
   outdir = os.path.dirname(output_file_path)
   os.makedirs(outdir, exist_ok=True)
 
-  content = """
-  #include <rclcpp/rclcpp.hpp>
-  #include <rclcpp_lifecycle/lifecycle_node.hpp>
-  #include "performance_test/node.hpp"
-
-  """
-
+  content = ""
   content += get_include_paths(msgs, srvs, package)
-  content += get_sub_factory(msgs, package, "rclcpp::Node")
-  content += get_sub_factory(msgs, package, "rclcpp_lifecycle::LifecycleNode")
+  content += get_sub_factory(msgs, package)
+  content += get_pub_factory(msgs, package)
+  content += get_server_factory(srvs, package)
+  content += get_client_factory(srvs, package)
 
-  content += get_pub_factory(msgs, package, "rclcpp::Node")
-  content += get_pub_factory(msgs, package, "rclcpp_lifecycle::LifecycleNode")
-
-  content += get_server_factory(srvs, package, "rclcpp::Node")
-  content += get_server_factory(srvs, package, "rclcpp_lifecycle::LifecycleNode")
-
-  content += get_client_factory(srvs, package, "rclcpp::Node")
-  content += get_client_factory(srvs, package, "rclcpp_lifecycle::LifecycleNode")
+  # Remove wacky indentation
+  content = textwrap.dedent(content)
 
   def create(filename, content):
     if os.path.exists(filename):
@@ -409,8 +368,6 @@ def main():
     open(filename, 'w').write(content)
 
   create(output_file_path, content)
-
-
 
 if __name__ == "__main__":
   main()
